Revert "fix!: UnmarshalJSON is limited to the default Tree bug (#277)"
This reverts commit bb5e119c2e8571fc8eb7aff3b6bca4915ffb4b18.
